update
풀링 시스템 개선
기사 B버튼, X버튼 누를 시 처리

+++ Assets/Scripts/99_Json/SkillDatas.cs
@@ -98,7 +98,7 @@
     public class KnightBInputData : Skillable
     {
         public float BuffDuration;
-        public float BuffIncreaseDef;
+        public int BuffIncreaseDef;
         
     }
